Iteration One Complete
Baskets have been added to the website, where the user can add an item
to there basket items which will appear in a list.

Also two modules have been created for discounts such as, money off, or
percentage off.

Current bug with Entity Framework not accepting Guid for the basket id,
https://github.com/aspnet/EntityFramework/releases

This prevents hackers from accessing a users basket. This will be
changed to a normal int shortly.

I will tidy up the GUI once the Entity Frameworks bug has been fixed, or
change it to a normal id